Changed footnotes to block quote notes
Footnotes are not convenient for the web format IMO --- In print, they appear on the same page as the reference, whereas the page size for HTML is not limited, and so the user might need to scroll quite a bit back up to the main text.

Here, I have attempted to replace them with block-quotes. I think it looks quite nice, but not 100% sure whether we should go with this.